tests: replace keystone with llvm-mc firmware build pipeline
Retire the keystone runtime assembler from Qbox CPU tests in favour of
firmware binaries built at CMake-configure time with llvm-mc + ld.lld
+ llvm-objcopy. Each test ships a paired .S file (with .quad/.word
placeholders for runtime-patched constants like MMIO_ADDR / NUM_WRITES)
and the new CpuTestBenchBase::load_firmware_binary() helper patches
those placeholders and ptr_load()s the result into m_mem at runtime.
Deletes patch/keystone/ and the keystone CPM fetch; LLVM tools are
now a hard build requirement (covered by the install_dependencies.sh
update in the previous commit).

Build-side helpers in tests/qbox/CMakeLists.txt:

- qbox_build_firmware_bin(<name> <arch> <src_dir> [EXTRA_LD_ARGS ...])
  assembles <src_dir>/<name>.S, links with ld.lld using
  <src_dir>/<name>.ld (or the per-directory firmware.ld fallback),
  and objcopies to a raw .bin via a <name>-firmware custom target.

- qbox_attach_firmware(<target> <fw_name>) wires the build dependency
  and defines FIRMWARE_BIN_PATH for the test executable.

- qbox_add_cpu_test(<target> <timeout_s> <sources>...
                    [ARCH <arch>] [FIRMWARE <name>]
                    [FIRMWARE_SRC_DIR <dir>] [EXTRA_LD_ARGS ...])
  collapses the build + attach + add_test boilerplate into one call.
  By default the firmware base name is the stem (NAME_WE) of the
  first source file; FIRMWARE overrides it.

- qbox_add_cpu_<arch>_test(<target> <timeout_s> <source>.cc) per-arch
  thin wrappers in cpu/<arch>/CMakeLists.txt that pin ARCH=<arch>.
  This is what call sites use, so each test reads as one line.

Portability:

- On MSYS2/Windows, ld.lld defaults to the lld-link/PE-COFF driver and
  silently ignores -flavor gnu, leaving GNU-style options like -T
  unrecognised. We pass the correct ELF emulation per arch
  (aarch64elf / hexagonelf / elf32lriscv / elf64lriscv) via -m on
  WIN32; older Linux LLDs (e.g. Ubuntu's stock llvm-18) don't know
  every emulation name, so we leave -m off elsewhere.

- hexagon-smmu-stress-test-v2.s uses Hexagon V73 ops (tlbw, dmwait,
  dmlink) that older LLVM Hexagon backends don't recognise. macOS
  via Homebrew's llvm and MSYS2/Windows both ship a recent-enough
  backend; the firmware (and its test) are gated behind APPLE OR
  WIN32 (QQVPQSP-714).

- tests/CMakeLists.txt gates each tests/qbox/cpu/<arch>/ subdirectory
  on the corresponding architecture being present in LIBQEMU_TARGETS.

Naming convention applied across all cpu-test sources:

- Filenames use dashes only, never underscores; the hexagon dir uses
  hexagon- consistently rather than the previous hex- shortening.

- Each test owns its firmware: aarch64 reset-test was previously
  shared by reset-test-system and reset-test-cpu; the .S is now
  duplicated so each has its own. hexagon-smmu-firmware was renamed
  to hexagon-smmu-stress-test-v2 to match its driver.

- After these renames every qbox_add_cpu_<arch>_test call has
  firmware == source-stem, so qbox_add_cpu_test derives the firmware
  base from the first source's NAME_WE and call sites just read
  (target, timeout, source).

Covers all aarch64, hexagon, and riscv32 CPU tests previously using
set_firmware() (including the multi-block smmu-router-stress-test-v2
layout via .org), and the standalone hexagon load-store-test harness.
