--- patch/keystone/0001-CMake-Add-support-for-modern-CMake.patch
@@ -1,120 +1,0 @@
-From f7cf646a865d469f9abcdc49ae36d1f1b1077af4 Mon Sep 17 00:00:00 2001
-From: Jerome Haxhiaj <jhaxhiaj@qti.qualcomm.com>
-Date: Wed, 10 Dec 2025 09:20:21 +0100
-Subject: [PATCH 1/2] CMake: Add support for modern CMake
-
-Signed-off-by: Jerome Haxhiaj <jhaxhiaj@qti.qualcomm.com>
----
- CMakeLists.txt        | 35 ++++++++++-------------------------
- kstool/CMakeLists.txt |  2 +-
- llvm/CMakeLists.txt   | 15 +--------------
- 3 files changed, 12 insertions(+), 40 deletions(-)
-
-diff --git a/CMakeLists.txt b/CMakeLists.txt
-index eb52c19..32f1bff 100644
---- a/CMakeLists.txt
-+++ b/CMakeLists.txt
-@@ -1,7 +1,7 @@
- # Keystone Assembler Engine (www.keystone-engine.org)
- # By Nguyen Anh Quynh, 2016
- 
--cmake_minimum_required(VERSION 2.8.7)
-+cmake_minimum_required(VERSION 3.10.0)
- project(keystone)
- 
- set(KEYSTONE_VERSION_MAJOR 0)
-@@ -14,23 +14,6 @@ if (NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
-   set(CMAKE_BUILD_TYPE "Debug")
- endif()
- 
--if (POLICY CMP0022)
--  cmake_policy(SET CMP0022 NEW) # automatic when 2.8.12 is required
--endif()
--
--if (POLICY CMP0051)
--  # CMake 3.1 and higher include generator expressions of the form
--  # $<TARGETLIB:obj> in the SOURCES property.  These need to be
--  # stripped everywhere that access the SOURCES property, so we just
--  # defer to the OLD behavior of not including generator expressions
--  # in the output for now.
--  cmake_policy(SET CMP0051 OLD)
--endif()
--
--if (POLICY CMP0063)
--  set(CMAKE_POLICY_DEFAULT_CMP0063 NEW) # automatic when 3.3.2 is required
--endif()
--
- if (CMAKE_VERSION VERSION_LESS 3.1.20141117)
-   set(cmake_3_2_USES_TERMINAL)
- else()
-@@ -107,12 +90,14 @@ INSTALL(FILES "${PKG_CONFIG_FILE_PATH}"
-         DESTINATION lib${LLVM_LIBDIR_SUFFIX}/pkgconfig)
- 
- # uninstall target
--configure_file(
--    "${CMAKE_CURRENT_SOURCE_DIR}/CMakeUninstall.in"
--    "${CMAKE_CURRENT_BINARY_DIR}/CMakeUninstall.cmake"
--    IMMEDIATE @ONLY)
--
--add_custom_target(uninstall
--    COMMAND ${CMAKE_COMMAND} -P ${CMAKE_CURRENT_BINARY_DIR}/CMakeUninstall.cmake)
-+if(NOT TARGET uninstall)
-+    configure_file(
-+        "${CMAKE_CURRENT_SOURCE_DIR}/CMakeUninstall.in"
-+        "${CMAKE_CURRENT_BINARY_DIR}/CMakeUninstall.cmake"
-+        IMMEDIATE @ONLY)
-+
-+    add_custom_target(uninstall
-+        COMMAND ${CMAKE_COMMAND} -P ${CMAKE_CURRENT_BINARY_DIR}/CMakeUninstall.cmake)
-+endif()
- 
- add_subdirectory(suite/fuzz)
-\ No newline at end of file
-diff --git a/kstool/CMakeLists.txt b/kstool/CMakeLists.txt
-index d28da48..2511382 100644
---- a/kstool/CMakeLists.txt
-+++ b/kstool/CMakeLists.txt
-@@ -1,7 +1,7 @@
- # Kstool for Keystone assembler engine.
- # By Nguyen Anh Quynh, 2016
- 
--cmake_minimum_required(VERSION 2.8)
-+cmake_minimum_required(VERSION 3.10.0)
- 
- project(kstool)
- 
-diff --git a/llvm/CMakeLists.txt b/llvm/CMakeLists.txt
-index e324ef3..8ef8e27 100644
---- a/llvm/CMakeLists.txt
-+++ b/llvm/CMakeLists.txt
-@@ -1,6 +1,6 @@
- # See docs/CMake.html for instructions about how to build LLVM with CMake.
- 
--cmake_minimum_required(VERSION 2.8.7)
-+cmake_minimum_required(VERSION 3.10.0)
- 
- set(LLVM_INSTALL_TOOLCHAIN_ONLY ON)
- 
-@@ -9,19 +9,6 @@ if (NOT CMAKE_BUILD_TYPE AND NOT CMAKE_CONFIGURATION_TYPES)
-   set(CMAKE_BUILD_TYPE "Debug")
- endif()
- 
--if(POLICY CMP0022)
--  cmake_policy(SET CMP0022 NEW) # automatic when 2.8.12 is required
--endif()
--
--if (POLICY CMP0051)
--  # CMake 3.1 and higher include generator expressions of the form
--  # $<TARGETLIB:obj> in the SOURCES property.  These need to be
--  # stripped everywhere that access the SOURCES property, so we just
--  # defer to the OLD behavior of not including generator expressions
--  # in the output for now.
--  cmake_policy(SET CMP0051 OLD)
--endif()
--
- if(CMAKE_VERSION VERSION_LESS 3.1.20141117)
-   set(cmake_3_2_USES_TERMINAL)
- else()
--- 
-2.54.0
-
