--- patch/keystone/0002-Honor-parent-project-s-CMAKE_RUNTIME_OUTPUT_DIRECTOR.patch
@@ -1,47 +1,0 @@
-From bc790fabb24e6ed9a6d284f280a8876788850431 Mon Sep 17 00:00:00 2001
-From: Jerome Haxhiaj <jhaxhiaj@qti.qualcomm.com>
-Date: Mon, 4 May 2026 21:56:50 +0200
-Subject: [PATCH 2/2] Honor parent project's CMAKE_RUNTIME_OUTPUT_DIRECTORY
-
-LLVM_RUNTIME_OUTPUT_INTDIR is what set_output_directory() actually uses
-to place shared libraries.  When a parent project has already set
-CMAKE_RUNTIME_OUTPUT_DIRECTORY, use that value so DLLs land alongside
-the rest of the build tree instead of in keystone's private bin/ dir.
-
-Signed-off-by: Jerome Haxhiaj <jhaxhiaj@qti.qualcomm.com>
----
- llvm/CMakeLists.txt | 11 +++++++++--
- 1 file changed, 9 insertions(+), 2 deletions(-)
-
-diff --git a/llvm/CMakeLists.txt b/llvm/CMakeLists.txt
-index 8ef8e27..f90e7e5 100644
---- a/llvm/CMakeLists.txt
-+++ b/llvm/CMakeLists.txt
-@@ -136,7 +136,11 @@ endif()
- set(LLVM_LIBDIR_SUFFIX "" CACHE STRING "Define suffix of library directory name (32/64)" )
- 
- # They are used as destination of target generators.
--set(LLVM_RUNTIME_OUTPUT_INTDIR ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}/bin)
-+if(CMAKE_RUNTIME_OUTPUT_DIRECTORY)
-+  set(LLVM_RUNTIME_OUTPUT_INTDIR ${CMAKE_RUNTIME_OUTPUT_DIRECTORY})
-+else()
-+  set(LLVM_RUNTIME_OUTPUT_INTDIR ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}/bin)
-+endif()
- set(LLVM_LIBRARY_OUTPUT_INTDIR ${CMAKE_CURRENT_BINARY_DIR}/${CMAKE_CFG_INTDIR}/lib${LLVM_LIBDIR_SUFFIX})
- if(WIN32 OR CYGWIN)
-   # DLL platform -- put DLLs into bin.
-@@ -436,7 +440,10 @@ configure_file(
-   ${LLVM_INCLUDE_DIR}/llvm/Support/DataTypes.h)
- 
- # They are not referenced. See set_output_directory().
--set( CMAKE_RUNTIME_OUTPUT_DIRECTORY ${LLVM_BINARY_DIR}/bin )
-+if (NOT CMAKE_RUNTIME_OUTPUT_DIRECTORY)
-+  set( CMAKE_RUNTIME_OUTPUT_DIRECTORY ${LLVM_BINARY_DIR}/bin )
-+endif (NOT CMAKE_RUNTIME_OUTPUT_DIRECTORY)
-+
- set( CMAKE_LIBRARY_OUTPUT_DIRECTORY ${LLVM_BINARY_DIR}/lib${LLVM_LIBDIR_SUFFIX} )
- set( CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${LLVM_BINARY_DIR}/lib${LLVM_LIBDIR_SUFFIX} )
- 
--- 
-2.54.0
-
