+++ tests/qbox/cpu/CMakeLists.txt
@@ -1,3 +1,9 @@
-add_subdirectory(aarch64)
-add_subdirectory(hexagon)
-add_subdirectory(riscv32)
+if("aarch64" IN_LIST LIBQEMU_TARGETS)
+    add_subdirectory(aarch64)
+endif()
+if("hexagon" IN_LIST LIBQEMU_TARGETS)
+    add_subdirectory(hexagon)
+endif()
+if("riscv32" IN_LIST LIBQEMU_TARGETS)
+    add_subdirectory(riscv32)
+endif()
