+++ tests/qbox/cpu/aarch64/firmware.ld
@@ -1,0 +1,12 @@
+ENTRY(_start)
+
+SECTIONS
+{
+    . = 0x0;
+    .text : {
+        *(.text)
+    }
+    .data : {
+        *(.data)
+    }
+}
