+++ tests/qbox/cpu/hexagon/firmware.ld
@@ -1,0 +1,12 @@
+ENTRY(boot_start)
+
+SECTIONS
+{
+    . = 0x0;
+    .text : {
+        *(.text)
+    }
+    .data : {
+        *(.data)
+    }
+}
