+++ tests/qbox/cpu/hexagon/hexagon-smmu-stress-test-v2.cc
@@ -7,25 +7,30 @@
  */
 
 #include <systemc>
+#include <array>
 #include <cstdio>
+#include <memory>
 #include <vector>
 #include <queue>
+#include <thread>
 #include <scp/report.h>
 #include <cci/utils/broker.h>
 #include <libgsutils.h>
 #include <cciutils.h>
 #include <argparser.h>
-#include <keystone/keystone.h>
 #include <tlm_utils/tlm_quantumkeeper.h>
 #include "cci/cfg/cci_broker_if.h"
+#include "sysc/kernel/sc_time.h"
 #include "test/cpu.h"
 #include "test/tester/mmio.h"
-#include "cortex-a53.h"
-#include "qemu-instance.h"
+#include <hexagon.h>
+#include <hexagon_globalreg.h>
+#include <qemu-instance.h>
 #include <router.h>
 #include <gs_memory.h>
 #include <smmu500.h>
 #include <tlm_utils/simple_initiator_socket.h>
+#include <tlm-extensions/pathid_extension.h>
 #include <pass.h>
 
 /*
@@ -136,6 +141,7 @@ public:
     static constexpr uint64_t REG_ITERATIONS = 0x20; // CPU reads: current iteration count
     static constexpr uint64_t REG_DEBUG = 0x28;      // CPU writes: debug messages
     static constexpr uint64_t REG_DEBUG_DATA = 0x30; // CPU writes: debug messages
+    static constexpr uint64_t REG_CPU_ID = 0x48;     // CPU reads: global CPU ID (from pathid)
 
     // Request types
     enum RequestType { NONE = 0, FILL_REQUEST = 1, CHECK_REQUEST = 2 };
@@ -156,7 +162,7 @@ public:
     tlm_utils::simple_target_socket<SMMUTesterController> socket;
 
     // Reference to parent test bench for SMMU control
-    class CpuArmCortexA53SMMUStressTestV2* m_parent;
+    class CpuHexagonSMMUStressTestV2* m_parent;
 
     std::vector<CPUState> m_cpu_states;
     std::queue<uint32_t> m_available_regions;
@@ -191,13 +197,23 @@ public:
                          << " regions, target " << target_iterations << " iterations";
     }
 
-    void set_parent(class CpuArmCortexA53SMMUStressTestV2* parent) { m_parent = parent; }
+    void set_parent(class CpuHexagonSMMUStressTestV2* parent) { m_parent = parent; }
 
     virtual void b_transport(tlm::tlm_generic_payload& trans, sc_core::sc_time& delay)
     {
         uint64_t addr = trans.get_address();
-        uint32_t cpu_id = static_cast<uint32_t>(addr / REG_SIZE_PER_CPU);
-        uint64_t reg_offset = addr % REG_SIZE_PER_CPU;
+        uint64_t reg_offset = addr; // No per-CPU offset, just use address directly
+
+        // Extract CPU ID from pathid extension
+        gs::PathIDExtension* ext = nullptr;
+        trans.get_extension(ext);
+        sc_assert(ext);
+        // 0,1 , 4,5 , 8,9 is CPU_0, 2, 4, ....
+        // 2,3 , 6,7 , 10,11 is CPU_1, 3, 5, ....
+        uint32_t port_id = ext->at(1);
+        uint32_t cpu_id = port_id >> 1; // Extract global CPU ID from pathid
+
+        SCP_DEBUG(())("Came through port {}, giving CPU_ID {}", ext->at(1), cpu_id);
 
         if (cpu_id >= m_num_cpus) {
             trans.set_response_status(tlm::TLM_ADDRESS_ERROR_RESPONSE);
@@ -272,12 +288,27 @@ private:
             data = m_global_iterations;
             break;
 
+        case REG_CPU_ID:
+            data = cpu_id; // Return the global CPU ID from pathid
+            break;
+
         default:
             SCP_FATAL(()) << "CPU " << cpu_id << " read from unknown register 0x" << std::hex << reg_offset;
             break;
         }
 
-        *reinterpret_cast<uint64_t*>(trans.get_data_ptr()) = data;
+        // Handle both 32-bit and 64-bit reads
+        unsigned int data_length = trans.get_data_length();
+        if (data_length == 4) {
+            // 32-bit read (memw instruction)
+            *reinterpret_cast<uint32_t*>(trans.get_data_ptr()) = static_cast<uint32_t>(data & 0xFFFFFFFF);
+        } else if (data_length == 8) {
+            // 64-bit read (memd instruction)
+            *reinterpret_cast<uint64_t*>(trans.get_data_ptr()) = data;
+        } else {
+            SCP_FATAL(()) << "CPU " << cpu_id << " unsupported read size " << data_length << " bytes from register 0x"
+                          << std::hex << reg_offset;
+        }
     }
 
     /**
@@ -311,9 +342,6 @@ private:
                 // Configure SMMU mapping for this CPU to this region
                 configure_smmu_mapping(cpu_id, region);
 
-                // Mark as ready
-                cpu.status = READY;
-
                 SCP_INFO((TEST)) << "CPU_" << cpu_id << " assigned region " << region << " for filling";
             } else {
                 cpu.status = BUSY; // No regions available
@@ -331,9 +359,6 @@ private:
                 // Configure SMMU mapping for this CPU to this region
                 configure_smmu_mapping(cpu_id, region);
 
-                // Mark as ready
-                cpu.status = READY;
-
                 SCP_INFO((TEST)) << "CPU_" << cpu_id << " assigned region " << region << " for checking";
             } else {
                 cpu.status = BUSY; // No regions to check
@@ -342,7 +367,7 @@ private:
         }
     }
 
-    // Definition moved after CpuArmCortexA53SMMUStressTestV2 below
+    // Definition moved after CpuHexagonSMMUStressTestV2
 
     void handle_debug(uint32_t cpu_id, uint64_t data)
     {
@@ -408,8 +433,8 @@ private:
 };
 
 /**
- * @class CpuArmCortexA53SMMUStressTestV2
- * @brief A SystemC test bench for stressing the SMMUv2 with multiple CPUs.
+ * @class CpuHexagonSMMUStressTestV2
+ * @brief A SystemC test bench for stressing the SMMUv2 with multiple Hexagon CPUs.
  *
  * This test implements a sophisticated, tester-controlled architecture to
  * stress the SMMU-500. Unlike traditional approaches where CPUs manage their
@@ -446,10 +471,10 @@ private:
  * SMMU, capable of handling multiple CPUs concurrently without data corruption
  * or synchronization failures.
  */
-class CpuArmCortexA53SMMUStressTestV2 : public TestBench, public CpuTesterCallbackIface
+class CpuHexagonSMMUStressTestV2 : public TestBench, public CpuTesterCallbackIface
 {
 public:
-    static constexpr uint16_t MAX_ITERATIONS = 0x7ff; // Limited to 16-bit for ARM64 movz instruction
+    static constexpr uint16_t MAX_ITERATIONS = 0x7fff; // Limited to 16-bit for ARM64 movz instruction
     static constexpr uint64_t PATTERN_SIZE = 16;
     static constexpr unsigned BOUNDARY_BYTES = 80;
     sc_core::sc_time TEST_DURATION = sc_core::sc_time(30, sc_core::SC_SEC);
@@ -468,7 +493,7 @@ public:
     static constexpr size_t MAIN_MEM_SIZE = 256 * 1024 * 1024; // 256MB for main memory
 
     // Virtual and physical memory regions - Use high virtual address >32-bit
-    static constexpr uint64_t VIRTUAL_TEST_ADDR = 0x80000000ULL; // 12GB virtual address
+    static constexpr uint64_t VIRTUAL_TEST_ADDR = 0x300000000ULL; // 12GB virtual address
     static constexpr uint64_t
         PAGE_TABLE_BASE = 0x10000000; // Page tables: 0x10000000 - 0x1007FFFF (512KB for up to 32 CPUs)
     static constexpr uint64_t REGION_BASE = 0x10080000; // Test regions: 0x10080000+ in 8KB blocks (after page tables)
@@ -495,8 +520,11 @@ protected:
     QemuInstanceManager m_inst_manager;
     QemuInstance m_inst_a;
     QemuInstance m_inst_b;
+    hexagon_globalreg m_hex_gregs_a;
+    hexagon_globalreg m_hex_gregs_b;
     bool ab = false;
-    sc_core::sc_vector<cpu_arm_cortexA53> m_cpus;
+    sc_core::sc_vector<qemu_cpu_hexagon> m_cpus;
+    std::array<std::unique_ptr<hexagon_tlb>, 2> m_tlbs;
 
     // Memory components
     gs::gs_memory<> m_mem;
@@ -517,6 +545,8 @@ protected:
 
     uint32_t m_num_regions;
     std::vector<uint32_t> m_cpu_to_region;
+
+    tlm_utils::simple_target_socket<SMMUTesterController> invalidate_dummy_socket;
 
 public:
     // Memory accessor for proper routing - made public for tester access
@@ -570,62 +600,49 @@ private:
     void reconfigure_context_bank(uint32_t cb, uint64_t page_table_addr);
 
 protected:
-    void set_firmware(const char* assembly, uint64_t addr = 0)
+    void set_firmware_from_binary(const uint8_t* binary, size_t size, uint64_t addr = 0)
     {
-        ks_engine* ks;
-        ks_err err;
-        size_t size, count;
-        uint8_t* fw;
+        m_mem.load.ptr_load(const_cast<uint8_t*>(binary), addr, size);
+        SCP_INFO(()) << "Loaded " << size << " bytes of Hexagon firmware at 0x" << std::hex << addr;
+    }
 
-        err = ks_open(KS_ARCH_ARM64, KS_MODE_LITTLE_ENDIAN, &ks);
+    void dmi_invalidation_thread()
+    {
+        while (true) {
+            wait(sc_core::sc_time((rand() % 100) * 100, sc_core::SC_US)); // Very frequent
 
-        if (err != KS_ERR_OK) {
-            SCP_FATAL(()) << "Unable to initialize keystone";
+            // Invalidate random regions
+            //            uint64_t addr = REGION_BASE + (rand() % m_num_regions) * REGION_SIZE;
+            invalidate_dummy_socket->invalidate_direct_mem_ptr(REGION_BASE, 0xffffffff);
+            SCP_WARN(())("Invalidate All from (0x{:x})", REGION_BASE);
         }
-
-        if (ks_asm(ks, assembly, addr, &fw, &size, &count) != KS_ERR_OK || size == 0) {
-            std::cerr << assembly << "\n";
-            std::cerr << "errno: " << ks_errno(ks) << "\n";
-            std::cerr << "error: " << ks_strerror(ks_errno(ks)) << "\n";
-            SCP_INFO() << assembly;
-            TEST_FAIL("Unable to assemble the test firmware\n");
-        }
-
-        m_mem.load.ptr_load(fw, addr, size);
-
-        ks_free(fw);
-        ks_close(ks);
     }
 
 public:
-    CpuArmCortexA53SMMUStressTestV2(const sc_core::sc_module_name& n)
+    CpuHexagonSMMUStressTestV2(const sc_core::sc_module_name& n)
         : TestBench(n)
         , p_num_cpu("num_cpu", 4, "Number of CPUs to instantiate in the test")
         , p_quantum_ns("quantum_ns", 1000000, "Value of the global TLM-2.0 quantum in ns")
-        , m_inst_a("inst_a", &m_inst_manager, cpu_arm_cortexA53::ARCH)
-        , m_inst_b("inst_b", &m_inst_manager, cpu_arm_cortexA53::ARCH)
+        , m_inst_a("inst_a", &m_inst_manager, qemu_cpu_hexagon::ARCH)
+        , m_inst_b("inst_b", &m_inst_manager, qemu_cpu_hexagon::ARCH)
+        , m_hex_gregs_a("hex_gregs_a", &m_inst_a)
+        , m_hex_gregs_b("hex_gregs_b", &m_inst_b)
         , m_cpus("cpu", p_num_cpu,
                  [this](const char* n, int i) {
                      ab = !ab;
-                     return new cpu_arm_cortexA53(n, ab ? m_inst_a : m_inst_b);
+                     return new qemu_cpu_hexagon(n, ab ? m_inst_a : m_inst_b);
                  })
         , m_mem("mem", MEM_SIZE)
         , m_main_mem("main_mem", MAIN_MEM_SIZE)
         , m_smmu("smmu")
         , m_global_router("global_router")
         , m_tester_controller("tester_controller", p_num_cpu.get_value(),
-                              std::max(3u, static_cast<uint32_t>(p_num_cpu.get_value() * 3)), MAX_ITERATIONS)
+                              std::max(3u, static_cast<uint32_t>(p_num_cpu.get_value() * 3)),
+                              MAX_ITERATIONS / p_num_cpu)
         , m_memory_accessor("memory_accessor")
+        , invalidate_dummy_socket("dummy_invalidate_socket")
     {
         using tlm_utils::tlm_quantumkeeper;
-
-        // Set up ARM-specific CPU configuration
-        int i = 0;
-        for (cpu_arm_cortexA53& cpu : m_cpus) {
-            cpu.p_mp_affinity = i++;
-            cpu.p_has_el3 = false;
-            cpu.p_has_el2 = false; // Un-needed and unavailable for HVF/KVM
-        }
 
         sc_core::sc_time global_quantum(p_quantum_ns, sc_core::SC_NS);
         tlm_quantumkeeper::set_global_quantum(global_quantum);
@@ -652,6 +669,18 @@ public:
         uint32_t num_cpus = p_num_cpu.get_value();
         m_pass_identity.resize(num_cpus);
         m_tbus_high_va.resize(num_cpus);
+        // Hexagon HW has one TLB per core shared by all HW threads; here each
+        // QEMU instance plays the role of a core, so we create exactly two
+        // TLBs and link every CPU to the one on its instance.
+        m_tlbs[0] = std::make_unique<hexagon_tlb>("cpu_tlb_a", m_inst_a);
+        m_tlbs[1] = std::make_unique<hexagon_tlb>("cpu_tlb_b", m_inst_b);
+        for (auto& tlb : m_tlbs) {
+            tlb->p_num_entries = 256;
+            tlb->p_dma_entries = 16;
+        }
+        for (uint32_t i = 0; i < num_cpus; ++i) {
+            m_cpus[i].set_hex_tlb(m_tlbs[i % 2].get());
+        }
 
         for (uint32_t i = 0; i < num_cpus; ++i) {
             // Identity pass-through for each CPU - bypasses SMMU for identity traffic
@@ -683,6 +712,9 @@ public:
         // Initialize CPU to region mapping
         m_cpu_to_region.resize(p_num_cpu.get_value(), 0xFFFFFFFF);
 
+        m_hex_gregs_a.p_hexagon_start_addr = BOOT_ADDR;
+        m_hex_gregs_b.p_hexagon_start_addr = BOOT_ADDR;
+
         // Set up tester controller parent reference
         m_tester_controller.set_parent(this);
 
@@ -699,9 +731,10 @@ public:
             m_cpu_routers[i]->add_target(m_tbus_high_va[i]->upstream_socket, VIRTUAL_TEST_ADDR, 0x10000000ULL);
 
             SCP_INFO(()) << "🔍 ROUTING DEBUG: CPU " << i << " -> CPU_Router_" << i;
-            SCP_INFO(()) << "  - Identity range [0x0 - 0x10000000] -> Pass_Identity_" << i << " (bypassing SMMU)";
-            SCP_INFO(()) << "  - High VA range [0x300000000 - 0x400000000] -> High_VA_TBU_" << i << " (StreamID "
-                         << (i + 1) << ")";
+            SCP_INFO(()) << "  - Identity range [0x0 - 0x" << std::hex << VIRTUAL_TEST_ADDR << "] -> Pass_Identity_"
+                         << i << " (bypassing SMMU)";
+            SCP_INFO(()) << "  - High VA range [0x" << std::hex << VIRTUAL_TEST_ADDR << " - end] -> High_VA_TBU_" << i
+                         << " (StreamID " << (i + 1) << ")";
             SCP_INFO(()) << "  - Identity pass name: " << m_pass_identity[i]->name();
             SCP_INFO(()) << "  - High VA TBU name: " << m_tbus_high_va[i]->name();
         }
@@ -732,12 +765,17 @@ public:
         SCP_INFO(()) << "  TESTER: 0x" << std::hex << TESTER_ADDR;
         SCP_INFO(()) << "  VIRTUAL_TEST: 0x" << std::hex << VIRTUAL_TEST_ADDR;
 
-        generate_tester_controlled_firmware();
+        SCP_INFO(()) << "🔥 ABOUT TO CALL load_hexagon_firmware()";
+        load_hexagon_firmware();
+        SCP_INFO(()) << "🔥 AFTER CALLING load_hexagon_firmware()";
 
         SC_THREAD(configure_test);
+
+        m_global_router.add_target(invalidate_dummy_socket, 0, 0);
+        SC_THREAD(dmi_invalidation_thread);
     }
 
-    virtual ~CpuArmCortexA53SMMUStressTestV2()
+    virtual ~CpuHexagonSMMUStressTestV2()
     {
         for (auto* pass : m_pass_identity) {
             delete pass;
@@ -748,6 +786,30 @@ public:
         }
         for (auto* router : m_cpu_routers) {
             delete router;
+        }
+    }
+
+    void before_end_of_elaboration() override
+    {
+        TestBench::before_end_of_elaboration();
+
+        m_hex_gregs_a.before_end_of_elaboration();
+        m_hex_gregs_b.before_end_of_elaboration();
+
+        qemu::Device hex_gregs_a_dev = m_hex_gregs_a.get_qemu_dev();
+        qemu::Device hex_gregs_b_dev = m_hex_gregs_b.get_qemu_dev();
+
+        for (int i = 0; i < m_cpus.size(); i++) {
+            auto& cpu = m_cpus[i];
+            cpu.before_end_of_elaboration();
+            cpu.p_vtcm_base_addr = REGION_BASE;
+            cpu.p_vtcm_size_kb = REGION_SIZE / 1024;
+            cpu.p_hexagon_num_threads = (m_cpus.size() + 1) / 2;
+            qemu::Device cpu_dev = cpu.get_qemu_dev();
+
+            // Link to appropriate global regs based on which instance
+            bool uses_inst_a = (i % 2 == 0);
+            cpu_dev.set_prop_link("global-regs", uses_inst_a ? hex_gregs_a_dev : hex_gregs_b_dev);
         }
     }
 
@@ -977,19 +1039,18 @@ public:
     }
     void clear_region_pattern(uint32_t region_id)
     {
-        uint64_t physical_addr = CpuArmCortexA53SMMUStressTestV2::REGION_BASE +
-                                 (region_id * CpuArmCortexA53SMMUStressTestV2::REGION_SIZE);
+        uint64_t physical_addr = CpuHexagonSMMUStressTestV2::REGION_BASE +
+                                 (region_id * CpuHexagonSMMUStressTestV2::REGION_SIZE);
 
         SCP_INFO(()) << "Clearing region " << region_id << " (2-PAGE) at physical address 0x" << std::hex
                      << physical_addr;
 
-        const uint32_t words_to_clear = CpuArmCortexA53SMMUStressTestV2::BOUNDARY_BYTES / 8;
-        const uint32_t num_pages = CpuArmCortexA53SMMUStressTestV2::REGION_SIZE /
-                                   CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE;
+        const uint32_t words_to_clear = CpuHexagonSMMUStressTestV2::BOUNDARY_BYTES / 8;
+        const uint32_t num_pages = CpuHexagonSMMUStressTestV2::REGION_SIZE / CpuHexagonSMMUStressTestV2::PAGE_SIZE;
 
         // Loop over all pages in the region (2 pages for 8KB regions)
         for (uint32_t page_num = 0; page_num < num_pages; ++page_num) {
-            uint64_t page_physical_addr = physical_addr + (page_num * CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE);
+            uint64_t page_physical_addr = physical_addr + (page_num * CpuHexagonSMMUStressTestV2::PAGE_SIZE);
 
             SCP_INFO(()) << "  - Clearing page " << page_num << " at PA=0x" << std::hex << page_physical_addr;
 
@@ -1001,7 +1062,7 @@ public:
             // Clear the end of this page
             for (uint32_t i = 0; i < words_to_clear; ++i) {
                 write_memory_64(
-                    page_physical_addr + CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE - (words_to_clear * 8) + (i * 8),
+                    page_physical_addr + CpuHexagonSMMUStressTestV2::PAGE_SIZE - (words_to_clear * 8) + (i * 8),
                     0xBEEF0000 + i);
             }
         }
@@ -1133,245 +1194,29 @@ public:
         m_cpu_to_region[cpu] = 0xFFFFFFFF;
     }
 
-    void generate_tester_controlled_firmware()
+    void load_hexagon_firmware()
     {
-        SCP_INFO(()) << "Generating tester-controlled firmware with proper layout";
+        SCP_INFO(()) << "🔥 LOADING HEXAGON FIRMWARE from file: " << FIRMWARE_BIN_PATH;
 
-        // 1. BOOT LOADER at 0x0 - Just jumps to main firmware
-        static constexpr const char* BOOT_LOADER = R"(
-            // BOOT LOADER at 0x0
-            boot_start:
-                // Jump to main firmware at 0x1000
-                movz x0, #0x%04)" PRIx32 R"(  // Main firmware address (0x1000)
-                br x0                         // Jump to main firmware
-        )";
+        // Read the firmware from the compiled binary file
+        std::ifstream firmware_file(FIRMWARE_BIN_PATH, std::ios::binary | std::ios::ate);
+        if (!firmware_file) {
+            SCP_FATAL(()) << "Failed to open firmware file: " << FIRMWARE_BIN_PATH;
+            return;
+        }
 
-        char boot_buf[strlen(BOOT_LOADER) + 1000];
-        std::snprintf(boot_buf, sizeof(boot_buf), BOOT_LOADER, static_cast<uint32_t>(MAIN_FIRMWARE_ADDR));
-        set_firmware(boot_buf, BOOT_ADDR);
+        std::streamsize firmware_size = firmware_file.tellg();
+        firmware_file.seekg(0, std::ios::beg);
 
-        // 2. DIAGNOSTIC HANDLER at 0x200 - Error reporting
-        static constexpr const char* DIAGNOSTIC_HANDLER = R"(
-            // DIAGNOSTIC ERROR HANDLER at 0x200
-            diagnostic_error:
-                // Get CPU ID
-                mrs x0, mpidr_el1
-                and x0, x0, #0xff
-                
-                // Calculate tester base address for this CPU
-                ldr x1, =0x%08)" PRIx64 R"(    // TESTER_ADDR
-                mov x2, #0x100                 // Register size per CPU
-                mul x2, x0, x2                 // CPU offset
-                add x1, x1, x2                // x1 = CPU-specific tester base
-                
-                // Send error message: 0xE000 + CPU_ID
-                mov x2, #0xE000
-                orr x2, x2, x0                // Error code + CPU ID
-                str x2, [x1, #0x28]           // Write to REG_DEBUG
-                
-                // Stop the test with error
-                mov x0, #1                    // Exit with error
-                hlt #0                        // Halt with error
-                
-            diagnostic_loop:
-                wfi
-                b diagnostic_loop
-        )";
+        std::vector<uint8_t> firmware_data(firmware_size);
+        if (!firmware_file.read(reinterpret_cast<char*>(firmware_data.data()), firmware_size)) {
+            SCP_FATAL(()) << "Failed to read firmware file: " << FIRMWARE_BIN_PATH;
+            return;
+        }
 
-        char diagnostic_buf[strlen(DIAGNOSTIC_HANDLER) + 1000];
-        std::snprintf(diagnostic_buf, sizeof(diagnostic_buf), DIAGNOSTIC_HANDLER, TESTER_ADDR);
-        set_firmware(diagnostic_buf, DIAGNOSTIC_ADDR);
-
-        // 3. MAIN FIRMWARE at 0x1000 - The actual test logic
-        static constexpr const char* MAIN_FIRMWARE = R"(
-            // MAIN FIRMWARE at 0x1000
-            main_start:
-                // Get CPU ID
-                mrs x0, mpidr_el1
-                and x0, x0, #0xff
-                mov x20, x0                    // Save CPU ID in x20
-
-                // Calculate tester base address for this CPU
-                ldr x1, =0x%08)" PRIx64 R"(    // TESTER_ADDR
-                mov x2, #0x100                 // Register size per CPU
-                mul x2, x20, x2                // CPU offset
-                add x21, x1, x2               // x21 = CPU-specific tester base
-
-                // Signal startup
-                mov x0, #0x1000
-                str x0, [x21, #0x28]          // Write to REG_DEBUG
-
-                // Initialize iteration counter
-                mov x22, #0                   // x22 = iteration counter
-
-            main_loop:
-                // Check if we've reached max iterations
-                mov x3, #%d                   // Max iterations
-                cmp x22, x3
-                b.ge test_complete
-
-                // Signal entering main loop
-                mov x0, #0x2000
-                orr x0, x0, x22               // Include iteration count
-                str x0, [x21, #0x28]          // Write to REG_DEBUG
-
-                // Try to get a region to fill
-                mov x0, #1                    // FILL_REQUEST
-                str x0, [x21, #0x00]          // Write to REG_REQUEST
-
-                // Poll for readiness
-            poll_fill:
-                mov x0, #0x3000               // Polling debug message
-                str x0, [x21, #0x28]          // Write to REG_DEBUG
-                
-                ldr x0, [x21, #0x08]          // Read REG_STATUS
-                cmp x0, #1                    // READY?
-                b.eq fill_ready
-                cmp x0, #0                    // BUSY?
-                b.eq try_check                // Try checking instead
-                b poll_fill
-
-            fill_ready:
-                // Get assigned region ID
-                ldr x23, [x21, #0x10]         // Read REG_REGION_ID
-                
-                // Signal starting work
-                mov x0, #0x4000
-                orr x0, x0, x23               // Include region ID
-                str x0, [x21, #0x28]          // Write to REG_DEBUG
-
-                // Fill the region at virtual address - DYNAMIC PAGE COUNT
-                ldr x3, =0x%016)" PRIx64 R"(   // VIRTUAL_TEST_ADDR (base)
-                mov x4, #%u                   // Boundary bytes to write at each boundary
-                lsr x4, x4, #3                // Convert bytes to 8-byte words
-                mov x26, #0                   // Page number (starts at 0)
-                mov x19, #%u                  // Number of pages in region (use x19, not x30)
-                
-            fill_page_loop:
-                cmp x26, x19                  // Loop over num_pages (using x19)
-                b.ge fill_done
-                
-                // Calculate page base address: base + (page_num * PAGE_SIZE)
-                mov x27, #1                   // Start with 1
-                lsl x27, x27, #12             // Shift left by 12 to get 0x1000 (4KB)
-                mul x28, x26, x27             // page_offset = page_num * PAGE_SIZE
-                add x29, x3, x28              // page_base = VIRTUAL_TEST_ADDR + page_offset
-                
-                // Call fill_boundary for the start of the page
-                mov x0, x29                   // Arg 1: base address
-                bl fill_boundary
-
-                // Calculate end boundary address
-                mov x6, #1
-                lsl x6, x6, #12               // x6 = 0x1000 (PAGE_SIZE)
-                sub x6, x6, x4, lsl #3        // x6 = 0x1000 - (words * 8)
-                add x0, x29, x6               // end_boundary_addr = page_base + offset
-                
-                // Call fill_boundary for the end of the page
-                bl fill_boundary
-
-            fill_page_next:
-                add x26, x26, #1              // Next page
-                b fill_page_loop
-
-            fill_done:
-                // Signal fill complete
-                mov x0, #1
-                str x0, [x21, #0x18]
-
-                add x22, x22, #1
-                b main_loop
-
-            try_check:
-                // No regions to check - just loop back
-                b main_loop
-
-            test_complete:
-                b end
-
-            end:
-                str x0, [x21, #0x30]
-                wfi
-                b end
-
-            // -------------------------------------------------------------
-            // fill_boundary function
-            //
-            // Fills a memory boundary with an enhanced, verifiable pattern.
-            // Pattern: CPU_ID | REGION_ID | PAGE_NUM | WORD_OFFSET
-            //
-            // Assumed Global Registers (read-only):
-            //   - x20: CPU ID
-            //   - x23: Region ID
-            //   - x26: Current Page Number
-            //   - x4:  Number of 8-byte words to write
-            //
-            // Arguments:
-            //   - x0:  Base address to start writing from
-            //
-            // Clobbered Registers (Temporaries):
-            //   - x5, x6, x7
-            // -------------------------------------------------------------
-            fill_boundary:
-                mov x5, #0                    // x5: Word offset, loop counter
-            fill_boundary_loop:
-                cmp x5, x4                    // Loop for `x4` words
-                b.ge fill_boundary_done
-
-                // --- Start Pattern Generation ---
-                mov x6, #0                    // x6: The final pattern register. Clear before use.
-
-                // 1. CPU ID (bits 63:32)
-                lsl x7, x20, #32              // Shift CPU ID into temp register x7
-                orr x6, x6, x7                // OR into pattern
-
-                // 2. Region ID (bits 31:16)
-                lsl x7, x23, #16              // Shift Region ID into temp register x7
-                orr x6, x6, x7                // OR into pattern
-
-                // 3. Page Number (bits 15:8)
-                lsl x7, x26, #8               // Shift Page Number into temp register x7
-                orr x6, x6, x7                // OR into pattern
-
-                // 4. Word Offset (bits 7:0)
-                orr x6, x6, x5                // OR Word Offset directly into pattern
-
-                // --- End Pattern Generation ---
-
-                // Calculate write address and store the pattern
-                lsl x7, x5, #3                // word_offset_in_bytes = word_offset * 8
-                add x7, x0, x7                // final_addr = base_addr + word_offset_in_bytes
-                
-                // DEBUG: Write the pattern and target address to the debug registers
-                //str x6, [x21, #0x28]
-                //str x7, [x21, #0x30]
-                
-                str x6, [x7]
-
-                add x5, x5, #1                // Increment word offset
-                b fill_boundary_loop
-
-            fill_boundary_done:
-                ret
-        )";
-
-        // Calculate number of pages per region
-        uint32_t num_pages = REGION_SIZE / PAGE_SIZE;
-
-        char main_buf[strlen(MAIN_FIRMWARE) + 1000];
-        std::snprintf(main_buf, sizeof(main_buf), MAIN_FIRMWARE,
-                      TESTER_ADDR,                           // %1: TESTER_ADDR
-                      MAX_ITERATIONS,                        // %2: MAX_ITERATIONS
-                      VIRTUAL_TEST_ADDR,                     // %3: VIRTUAL_TEST_ADDR (fill)
-                      static_cast<unsigned>(BOUNDARY_BYTES), // %4: boundary bytes (fill)
-                      num_pages);                            // %5: number of pages per region
-
-        set_firmware(main_buf, MAIN_FIRMWARE_ADDR);
-
-        SCP_INFO(()) << "Firmware layout completed:";
-        SCP_INFO(()) << "  - Boot loader at 0x" << std::hex << BOOT_ADDR;
-        SCP_INFO(()) << "  - Diagnostic handler at 0x" << std::hex << DIAGNOSTIC_ADDR;
-        SCP_INFO(()) << "  - Main firmware at 0x" << std::hex << MAIN_FIRMWARE_ADDR;
+        SCP_INFO(()) << "🔥 LOADED " << firmware_size << " bytes of Hexagon firmware, calling set_firmware_from_binary";
+        set_firmware_from_binary(firmware_data.data(), firmware_size, BOOT_ADDR);
+        SCP_INFO(()) << "🔥 FIRMWARE SET COMPLETE";
     }
 
     void write_smmu_register(uint32_t addr, uint32_t value)
@@ -1419,7 +1264,7 @@ public:
         write_smmu_register(tlbiall_addr, 0x0); // Any write invalidates all entries for this CB
 
         // Add delay to ensure TLB invalidation takes effect
-        wait(sc_core::sc_time(2, sc_core::SC_US));
+        // wait(sc_core::sc_time(2, sc_core::SC_US));
 
         SCP_INFO(()) << "✅ TLB INVALIDATION COMPLETE: CPU " << cpu << " TLB cleared";
     }
@@ -1435,13 +1280,13 @@ public:
     {
         SCP_WARN(()) << "SMMU Stress Test V2 completed"
                      << "\nFinal statistics:" << "\n  - Total iterations: " << m_tester_controller.m_global_iterations
-                     << "/" << MAX_ITERATIONS << "\n  - CPUs: " << p_num_cpu.get_value()
+                     << "/" << MAX_ITERATIONS / p_num_cpu << "\n  - CPUs: " << p_num_cpu.get_value()
                      << "\n  - Memory regions: " << m_num_regions << "\n  - Pattern size: " << PATTERN_SIZE
                      << " * 8 bytes";
     }
 };
 
-void CpuArmCortexA53SMMUStressTestV2::reconfigure_context_bank(uint32_t cb, uint64_t page_table_addr)
+void CpuHexagonSMMUStressTestV2::reconfigure_context_bank(uint32_t cb, uint64_t page_table_addr)
 {
     uint32_t cb_base = get_context_bank_base(cb);
 
@@ -1483,7 +1328,15 @@ void CpuArmCortexA53SMMUStressTestV2::reconfigure_context_bank(uint32_t cb, uint
 void SMMUTesterController::configure_smmu_mapping(uint32_t cpu_id, uint32_t region_id)
 {
     if (m_parent) {
-        m_parent->map_cpu_to_region(cpu_id, region_id);
+        // Run map_cpu_to_region in a separate thread
+        std::thread mapping_thread([this, cpu_id, region_id]() {
+            CPUState& cpu = m_cpu_states[cpu_id];
+            m_parent->map_cpu_to_region(cpu_id, region_id);
+            cpu.status = READY;
+        });
+
+        // Detach the thread to allow it to run independently
+        mapping_thread.detach();
     }
 }
 
@@ -1502,19 +1355,18 @@ bool SMMUTesterController::verify_region_pattern(uint32_t cpu_id, uint32_t regio
         return false;
     }
 
-    uint64_t physical_addr = CpuArmCortexA53SMMUStressTestV2::REGION_BASE +
-                             (region_id * CpuArmCortexA53SMMUStressTestV2::REGION_SIZE);
+    uint64_t physical_addr = CpuHexagonSMMUStressTestV2::REGION_BASE +
+                             (region_id * CpuHexagonSMMUStressTestV2::REGION_SIZE);
 
     SCP_INFO(()) << "Verifying region " << region_id << " with ENHANCED 2-PAGE pattern at PA=0x" << std::hex
                  << physical_addr;
 
     const uint32_t words_to_check = 10; // Check 10 words (80 bytes) at each boundary
-    const uint32_t num_pages = CpuArmCortexA53SMMUStressTestV2::REGION_SIZE /
-                               CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE;
+    const uint32_t num_pages = CpuHexagonSMMUStressTestV2::REGION_SIZE / CpuHexagonSMMUStressTestV2::PAGE_SIZE;
 
     // Loop over all pages in the region (2 pages for 8KB regions)
     for (uint32_t page_num = 0; page_num < num_pages; ++page_num) {
-        uint64_t page_physical_addr = physical_addr + (page_num * CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE);
+        uint64_t page_physical_addr = physical_addr + (page_num * CpuHexagonSMMUStressTestV2::PAGE_SIZE);
 
         SCP_INFO(()) << "  - Verifying page " << page_num << " at PA=0x" << std::hex << page_physical_addr;
 
@@ -1542,8 +1394,8 @@ bool SMMUTesterController::verify_region_pattern(uint32_t cpu_id, uint32_t regio
 
         // Check the end of this page
         for (uint32_t word = 0; word < words_to_check; ++word) {
-            uint64_t word_addr = page_physical_addr + CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE -
-                                 (words_to_check * 8) + (word * 8);
+            uint64_t word_addr = page_physical_addr + CpuHexagonSMMUStressTestV2::PAGE_SIZE - (words_to_check * 8) +
+                                 (word * 8);
             uint64_t read_value = m_parent->m_memory_accessor.read_memory(word_addr);
 
             uint32_t pattern_cpu = (read_value >> 32) & 0xFFFFFFFF;
@@ -1578,18 +1430,18 @@ void SMMUTesterController::clear_region_pattern(uint32_t region_id)
         return;
     }
 
-    uint64_t physical_addr = CpuArmCortexA53SMMUStressTestV2::REGION_BASE +
-                             (region_id * CpuArmCortexA53SMMUStressTestV2::REGION_SIZE);
+    uint64_t physical_addr = CpuHexagonSMMUStressTestV2::REGION_BASE +
+                             (region_id * CpuHexagonSMMUStressTestV2::REGION_SIZE);
 
     SCP_INFO(()) << "Clearing region " << region_id << " (2-PAGE) at physical address 0x" << std::hex << physical_addr;
 
-    const uint32_t words_to_clear = CpuArmCortexA53SMMUStressTestV2::BOUNDARY_BYTES / 8;
-    const uint32_t num_pages = CpuArmCortexA53SMMUStressTestV2::REGION_SIZE /
-                               CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE;
+    const uint32_t words_to_clear = CpuHexagonSMMUStressTestV2::BOUNDARY_BYTES / 8;
+    const uint32_t num_pages = CpuHexagonSMMUStressTestV2::REGION_SIZE /
+                               CpuHexagonSMMUStressTestV2::PAGE_SIZE;
 
     // Loop over all pages in the region (2 pages for 8KB regions)
     for (uint32_t page_num = 0; page_num < num_pages; ++page_num) {
-        uint64_t page_physical_addr = physical_addr + (page_num * CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE);
+        uint64_t page_physical_addr = physical_addr + (page_num * CpuHexagonSMMUStressTestV2::PAGE_SIZE);
 
         SCP_INFO(()) << "  - Clearing page " << page_num << " at PA=0x" << std::hex << page_physical_addr;
 
@@ -1601,7 +1453,7 @@ void SMMUTesterController::clear_region_pattern(uint32_t region_id)
         // Clear the end of this page
         for (uint32_t i = 0; i < words_to_clear; ++i) {
             m_parent->write_memory_64(
-                page_physical_addr + CpuArmCortexA53SMMUStressTestV2::PAGE_SIZE - (words_to_clear * 8) + (i * 8),
+                page_physical_addr + CpuHexagonSMMUStressTestV2::PAGE_SIZE - (words_to_clear * 8) + (i * 8),
                 0xBEEF0000 + i);
         }
     }
@@ -1609,7 +1461,7 @@ void SMMUTesterController::clear_region_pattern(uint32_t region_id)
     SCP_INFO(()) << "✅ Region " << region_id << " cleared (all " << num_pages << " pages)";
 }
 */
-/* ---- Implementation moved here to ensure CpuArmCortexA53SMMUStressTestV2 is a complete type ---- */
+/* ---- Implementation moved here to ensure CpuHexagonSMMUStressTestV2 is a complete type ---- */
 
 /**
  * @brief Handles a completion notification from a CPU.
@@ -1634,19 +1486,16 @@ void SMMUTesterController::handle_complete(uint32_t cpu_id, CompleteType complet
 
         SCP_INFO(()) << "CPU " << cpu_id << " completed filling region " << cpu.assigned_region;
 
-        uint64_t physical_addr = CpuArmCortexA53SMMUStressTestV2::REGION_BASE +
-                                 (cpu.assigned_region * CpuArmCortexA53SMMUStressTestV2::REGION_SIZE);
-        uint64_t first_word = m_parent->m_memory_accessor.read_memory(physical_addr);
-        uint32_t original_filler_cpu = (first_word >> 32) & 0xFFFFFFFF;
-
-        if (verify_region_pattern(original_filler_cpu, cpu.assigned_region)) {
-            SCP_INFO((TEST))("Region {:d} (for CPU_{:d}) verified by tester", cpu.assigned_region, original_filler_cpu);
+        // Use cpu_id from pathid (the CPU that's reporting completion) to verify the pattern
+        // The testbench knows which CPU was assigned to this region, so we verify against that CPU ID
+        if (verify_region_pattern(cpu_id, cpu.assigned_region)) {
+            SCP_INFO((TEST))("Region {:d} filled by CPU_{:d} verified successfully", cpu.assigned_region, cpu_id);
             // clear_region_pattern(cpu.assigned_region);
             m_available_regions.push(cpu.assigned_region);
             cpu.iteration_count++;
             m_global_iterations++;
         } else {
-            SCP_FATAL(()) << "Region " << cpu.assigned_region << " verification failed";
+            SCP_FATAL(()) << "Region " << cpu.assigned_region << " verification failed for CPU " << cpu_id;
             sc_core::sc_stop();
             return;
         }
@@ -1673,6 +1522,7 @@ void SMMUTesterController::handle_complete(uint32_t cpu_id, CompleteType complet
 
 int sc_main(int argc, char* argv[])
 {
+    std::srand(std::time({}));
     scp::init_logging(scp::LogConfig()
                           .fileInfoFrom(sc_core::SC_ERROR)
                           .logAsync(false)
@@ -1684,8 +1534,8 @@ int sc_main(int argc, char* argv[])
     auto broker_h = m_broker.create_broker_handle(orig);
     ArgParser ap{ broker_h, argc, argv };
 
-    SCP_INFO("sc_main") << "Start SMMU Stress Test V2 - Tester Controlled";
-    CpuArmCortexA53SMMUStressTestV2 test_bench("test-bench");
+    SCP_INFO("sc_main") << "Start Hexagon SMMU Stress Test V2";
+    CpuHexagonSMMUStressTestV2 test_bench("test-bench");
 
     if ((gs::cci_get<int>(broker_h, "test-bench.num_cpu") > 8) ||
         gs::cci_get_d<bool>(broker_h, "test-bench.inst_a.icount", false)) {
