+++ tests/qbox/cpu/hexagon/load-store-test.cc
@@ -14,25 +14,6 @@
 class LoadStoreTest : public sc_core::sc_module, public MmioWriter, public MmioReader
 {
 private:
-    static constexpr const char* FIRMWARE = R"(
-        _start:
-            r0 = #0x%08)" PRIx32 R"(
-
-            //load and discard the result from CpuTesterMmio::MMIO_ADDR:
-            r2 = memw(r0)
-
-            //store a known value to CpuTesterMmio::MMIO_ADDR:
-            r3 = #0x0f0f0f0f
-            memw(r0) = r3
-        end:
-            // This instruction is not supported by keystone:
-            //   wait(r0)
-            // So we'll just bypass it and use the word we got from a reference
-            //   assembler:
-            .word 0x6440c000
-            jump end
-    )";
-
     static constexpr uint64_t MEM_ADDR = 0x0;
     static constexpr size_t MEM_SIZE = 256 * 1024;
     static constexpr uint32_t QUANTUM = 1000000;
@@ -70,9 +51,8 @@ public:
         router.add_target(memory.socket, MEM_ADDR, MEM_SIZE);
         router.add_initiator(cpu.socket);
 
-        char buf[1024];
-        std::snprintf(buf, sizeof(buf), FIRMWARE, static_cast<uint32_t>(MmioProbe::MMIO_ADDR));
-        load_firmware(memory, buf, MEM_ADDR);
+        load_firmware(memory, FIRMWARE_BIN_PATH, MEM_ADDR,
+                      std::initializer_list<uint32_t>{ static_cast<uint32_t>(MmioProbe::MMIO_ADDR) });
 
         mmio_probe.connect_to_mmio_writer(this);
         mmio_probe.connect_to_mmio_reader(this);
