+++ tests/qbox/cpu/riscv32/riscv32-reset.cc
@@ -48,56 +48,11 @@ class CpuRiscv32ResetGPIOTest : public CpuTestBench<cpu_riscv32, CpuTesterMmio>
     int reset_done;
     int time_elapsed_ms;
 
-    /*
-     * Load RISC-V 32-bit firmware from binary file compiled with LLVM tools.
-     * The binary contains the constants at the end that need to be patched.
-     */
-    void load_firmware_binary(uint32_t mmio_addr, uint32_t trigger_val, uint32_t done_val)
+    void load_reset_firmware(uint32_t mmio_addr, uint32_t trigger_val, uint32_t done_val)
     {
-        SCP_DEBUG(SCMOD) << "load_firmware_binary called with mmio_addr=0x" << std::hex << mmio_addr
-                         << ", trigger_val=" << std::dec << trigger_val << ", done_val=" << done_val;
-        // Load the compiled firmware binary
-        const char* firmware_path = FIRMWARE_BIN_PATH;
-        std::ifstream file(firmware_path, std::ios::binary | std::ios::ate);
-
-        if (!file.is_open()) {
-            SCP_FATAL(SCMOD) << "Failed to open RISC-V firmware file: " << firmware_path;
-            TEST_ASSERT(false);
-        }
-
-        std::streamsize size = file.tellg();
-        file.seekg(0, std::ios::beg);
-
-        std::vector<uint8_t> firmware_data(size);
-        if (!file.read(reinterpret_cast<char*>(firmware_data.data()), size)) {
-            SCP_FATAL(SCMOD) << "Failed to read RISC-V firmware file: " << firmware_path;
-            TEST_ASSERT(false);
-        }
-
-        // Patch the constants in the binary (last 12 bytes: mmio_addr, trigger_val, done_val)
-        TEST_ASSERT(size >= 12);
-        uint32_t* mmio_addr_ptr = reinterpret_cast<uint32_t*>(firmware_data.data() + size - 12);
-        uint32_t* trigger_val_ptr = reinterpret_cast<uint32_t*>(firmware_data.data() + size - 8);
-        uint32_t* done_val_ptr = reinterpret_cast<uint32_t*>(firmware_data.data() + size - 4);
-
-        *mmio_addr_ptr = mmio_addr;
-        *trigger_val_ptr = trigger_val;
-        *done_val_ptr = done_val;
-
-        // Log firmware being loaded (for verification during testing)
-        if (trigger_val == RESET_TRIGGER) {
-            SCP_INFO(SCMOD) << "Loading initial RISC-V firmware with trigger_val=" << trigger_val << " from "
-                            << firmware_path;
-        } else {
-            SCP_INFO(SCMOD) << "Loading reset firmware with trigger_val=" << trigger_val << " from " << firmware_path;
-        }
-
-        // Load firmware directly into memory without keystone assembly
-        SCP_DEBUG(SCMOD) << "Loading RISC-V firmware at 0x" << std::hex << MEM_ADDR << " (MEM_ADDR), size=" << std::dec
-                         << size << " bytes";
-        m_mem.load.ptr_load(firmware_data.data(), MEM_ADDR, size);
-
-        SCP_DEBUG(SCMOD) << "Firmware loaded successfully";
+        SCP_INFO(SCMOD) << "Loading RISC-V firmware (trigger_val=" << trigger_val << ") from " << FIRMWARE_BIN_PATH;
+        load_firmware_binary(FIRMWARE_BIN_PATH, MEM_ADDR,
+                             std::initializer_list<uint32_t>{ mmio_addr, trigger_val, done_val });
     }
 
 public:
@@ -124,7 +79,7 @@ public:
         dont_initialize();
 
         // Load RISC-V 32-bit binary firmware compiled with LLVM tools
-        load_firmware_binary(static_cast<uint32_t>(CpuTesterMmio::MMIO_ADDR), RESET_TRIGGER, RESET_DONE);
+        load_reset_firmware(static_cast<uint32_t>(CpuTesterMmio::MMIO_ADDR), RESET_TRIGGER, RESET_DONE);
     }
 
     virtual ~CpuRiscv32ResetGPIOTest() {}
@@ -148,9 +103,9 @@ public:
              * Load the final firmware image, now the image will write to this address again
              * with a RESET_DONE value and then wait.
              */
-            SCP_DEBUG(SCMOD) << "About to call load_firmware_binary with RESET_DONE";
-            load_firmware_binary(static_cast<uint32_t>(CpuTesterMmio::MMIO_ADDR), RESET_DONE, RESET_DONE);
-            SCP_DEBUG(SCMOD) << "load_firmware_binary call completed";
+            SCP_DEBUG(SCMOD) << "About to call load_reset_firmware with RESET_DONE";
+            load_reset_firmware(static_cast<uint32_t>(CpuTesterMmio::MMIO_ADDR), RESET_DONE, RESET_DONE);
+            SCP_DEBUG(SCMOD) << "load_reset_firmware call completed";
 
             /*
              * Invalidate TB cache to ensure cached translations of old firmware are cleared
